+++ lib/runger/settings.rb
@@ -96,11 +96,11 @@ module Runger
       end
     end
 
-    # By default, use ANYWAY_ENV
-    self.current_environment = ENV["ANYWAY_ENV"]
+    # By default, use RUNGER_ENV
+    self.current_environment = ENV["RUNGER_ENV"]
 
     # By default, use local files only in development (that's the purpose if the local files)
-    self.use_local_files = (ENV["ANYWAY_ENV"] == "development" || ENV["RACK_ENV"] == "development" || ENV["RAILS_ENV"] == "development" || (defined?(Rails) && Rails.env.development?))
+    self.use_local_files = (ENV["RUNGER_ENV"] == "development" || ENV["RACK_ENV"] == "development" || ENV["RAILS_ENV"] == "development" || (defined?(Rails) && Rails.env.development?))
 
     # By default, consider configs are stored in the ./config folder
     self.default_config_path = ->(name) { "./config/#{name}.yml" }